requirements.txt
coverage
dvc
numpy
pandas
scikit-learn
pyyaml
seaborn
fastapi[standard]
uvicorn
psycopg2-binary
hvac
python-dotenv